requirements-dev.txt
# Development dependencies
black==23.12.1
flake8==7.0.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-flask==1.3.0
mypy==1.8.0
pre-commit==3.6.0

# Testing
factory-boy==3.3.0
faker==22.2.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Code quality
bandit==1.7.5
safety==2.3.5
isort==5.13.2

# Development tools
python-dotenv==1.0.0
watchdog==3.0.0